Camera implementation
So the idea of the camera is pretty tricky if u wasn't familiar with
this trick early (i wasn't).

But idea is that we should make a code which will make camera follow
player, right?

so our player spawn on your left but camera at the center right now.
Now we take a delta which is the (player.x - window.width / 2). And then
in tilemap we just substract this value(lets say -30) and get (tile.x
-(-30)) so we get +30 and thus we moveing all scene by +30 ti the right.